Move redux bindings from settings group to parent settings form component.
