+++ client/components/wcc-settings-form/index.js
@@ -2,18 +2,54 @@ import React, { PropTypes } from 'react';
 import WCCSettingsGroup from './settings-group';
 import notices from 'notices';
 import GlobalNotices from 'components/global-notices';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import * as FormActions from 'state/form/actions';
+import { successNotice, errorNotice } from 'state/notices/actions';
+import isString from 'lodash/isString';
+import { translate as __ } from 'lib/mixins/i18n';
 
-const WCCSettingsForm = ( { storeOptions, schema, layout, saveFormData } ) => {
+const WCCSettingsForm = ( {
+	storeOptions,
+	schema,
+	layout,
+	settings,
+	form,
+	saveFormData,
+	formActions,
+	noticeActions,
+} ) => {
+	const setIsSaving = ( value ) => formActions.setField( 'isSaving', value );
+	const setSuccess = ( value ) => {
+		formActions.setField( 'success', value );
+		if ( true === value ) {
+			noticeActions.successNotice( __( 'Your changes have been saved.' ), {
+				duration: 2250,
+			} );
+		}
+	};
+	const setError = ( value ) => {
+		formActions.setField( 'error', value );
+		if ( isString( value ) ) {
+			noticeActions.errorNotice( value, {
+				duration: 7000,
+			} );
+		}
+	}
+	const saveForm = () => saveFormData( setIsSaving, setSuccess, setError, settings );
 	return (
 		<div>
 			<GlobalNotices id="notices" notices={ notices.list } />
 			{ layout.map( ( group, idx ) => (
 				<WCCSettingsGroup
 					key={ idx }
-					group={ group }
-					schema={ schema }
-					storeOptions={ storeOptions }
-					saveFormData={ saveFormData }
+					{ ...{
+						group,
+						schema,
+						storeOptions,
+						saveForm,
+						form,
+					} }
 				/>
 			) ) }
 		</div>
@@ -27,4 +63,21 @@ WCCSettingsForm.propTypes = {
 	saveFormData: PropTypes.func.isRequired,
 };
 
-export default WCCSettingsForm;
+function mapStateToProps( state ) {
+	return {
+		settings: state.settings,
+		form: state.form,
+	};
+}
+
+function mapDispatchToProps( dispatch ) {
+	return {
+		formActions: bindActionCreators( FormActions, dispatch ),
+		noticeActions: bindActionCreators( { successNotice, errorNotice }, dispatch ),
+	};
+}
+
+export default connect(
+	mapStateToProps,
+	mapDispatchToProps
+)( WCCSettingsForm );
